Added MultiJson dependency.
Until now, the library used .to_json methods for converting a resource
or error message or params to the corresponding JSON representation.
This locks the library to the stdlib json gem. With the current changes
in place, the fastest JSON serializer present in the load path is used.

For example, if the require directive "require 'oj'" is present in the
load path, then that serializer is used even if "json_pure" or
"json/ext" is present in the same load path.